allow bridging flows to also output to controller
Bridging DLF and multicast flows allow including the controller via the
output action, so add support for setting the value when adding the
flow.

since this means that the flow configuration may not be static, we also
need to add support for updating the flow to dynamically enable or
disable reception by the controller.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>