fix OXM_TLV_EXPR_VRF value
According to ofagent the code is 0xffff0206, i.e. class = 0xffff [bits
31:16], type = 0x1 [bits 15:9], mask = 0x0 [bit 8] and length = 0x6
[bits 7:0].

So the VRF action has the length 6, not 2. Also while OF_OXM_BSN has a
VRF_MASK variant, OF_OXM_OFDPA does not, so remove the non-existing
match defintion.

This hasn't caused any issues since rofl-common fixes up the length
field [1][2], but it's better to be correct in the first place.

[1] https://github.com/bisdn/rofl-common/blob/main/src/rofl/common/openflow/coxmatch.h#L279
[2] https://github.com/bisdn/rofl-common/blob/main/src/rofl/common/openflow/coxmatch.h#L523

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>